pyproject.toml
[build-system]
requires = [
    "setuptools>=64.0.0",
    "scikit-build-core",
    "pybind11>=2.9.0",
    "wheel",
]
build-backend = "scikit_build_core.build"

[project]
name = "tady"
version = "0.1.0"
description = "Machine learning-based binary analysis toolkit"
readme = "README.md"
requires-python = "==3.12.*"
license = {file = "LICENSE"}
authors = [
    {name = "Tady Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "tqdm",
    "numpy",
    "hydra-core",
    "hydra-joblib-launcher",
    "jax[cuda12]",
    "flax[all]",
    "grain",
    "lief",
    "datasets",
    "wandb",
    "transformers",
    "orbax-checkpoint",
    "tensorflow-cpu",
    "tensorflow-serving-api",
    "numpy",
    "matplotlib",
    "quokka-project",
]

[project.urls]
Repository = "https://github.com/5c4lar/tady"

[tool.setuptools]
package-dir = {"" = "python"}

[tool.scikit-build]
cmake.build-type = "Release"
cmake.args = ["-DTADY_ENABLE_CPP=ON"]
install.strip = false

[dependency-groups]
dev = [
    "ipykernel",
]

[tool.setuptools.packages.find]
where = ["python"]
include = ["tady*"]
